# MindApi/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the Master Inventory Network Database (MIND) API. MIND provides automated allocation of technical attributes needed to create IP and Ethernet VPNs on the global Sigma network. MIND supports the 'Nova' services specfication which defines the collection of connectivity services supported by ENT. Go to https://thehub.thomsonreuters.com/docs/DOC-2193014 to learn more.",
    "version" : "1.0.0",
    "title" : "MIND API",
    "contact" : {
      "email" : "jonathan.beasley@thomsonreuters.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "raid.thomsonreuters.com",
  "basePath" : "/v1",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/tenant/{tenantId}/device" : {
      "get" : {
        "tags" : [ "tenant device" ],
        "summary" : "Find all devices for a given tenant",
        "description" : "Returns all infrastructure devices",
        "operationId" : "getAllTenantDevices",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "description" : "ID of the tenant",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/TenantDevice"
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "tenant device" ],
        "summary" : "Create a new tenant device",
        "description" : "",
        "operationId" : "addTenantDevice",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "description" : "ID of the tenant",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "device request object that generates a new device",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TenantDeviceRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/TenantDevice"
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/tenant/device/{deviceId}" : {
      "get" : {
        "tags" : [ "tenant device" ],
        "summary" : "Find device by ID",
        "description" : "Returns a single device",
        "operationId" : "getTenantDeviceById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "ID of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TenantDevice"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "tenant device" ],
        "summary" : "Update an existing device",
        "description" : "",
        "operationId" : "updateTenantDevice",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "ID of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "device update object that updates an existing device",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TenantDeviceUpdate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/TenantDevice"
              }
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "tenant device" ],
        "summary" : "Deletes a device",
        "description" : "",
        "operationId" : "deleteTenantDevice",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "ID of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/tenant/device/{deviceId}/port" : {
      "get" : {
        "tags" : [ "tenant device" ],
        "summary" : "Find all ports for a given tenant device",
        "description" : "Returns all ports for a given tenant device",
        "operationId" : "getPortsByTenantDevice",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "ID of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Port"
              }
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "tenant device" ],
        "summary" : "Create a new port",
        "operationId" : "createTenantPort",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "ID of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Created port object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/PortRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Port"
              }
            }
          },
          "400" : {
            "description" : "Validation error",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/tenant/device/port/{portId}" : {
      "get" : {
        "tags" : [ "tenant device" ],
        "summary" : "Find port by ID",
        "description" : "Returns a single port",
        "operationId" : "getTenantPortById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "portId",
          "in" : "path",
          "description" : "ID of the port",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Port"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "tenant device" ],
        "summary" : "Deletes a port",
        "description" : "",
        "operationId" : "deleteTenantPort",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "portId",
          "in" : "path",
          "description" : "ID of the port",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          },
          "404" : {
            "description" : "The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      },
      "example" : {
        "code" : 0,
        "type" : "type",
        "message" : "message"
      }
    },
...
  },
...
}